msm: camera: csiphy: Fix misc seccam integration issues
The current secure camera implementation suffers with a few
issues in terms of typecasting, error handling and in
populating the information in the right data structures when
the new SCM API is used on the targets without domain ID support.
In addition, we needed to be explicit in failing to acquire
if dual IFE is required in secure camera use cases as it is
not a supported feature as of now.

CRs-Fixed: 3317248

+++ drivers/cam_isp/isp_hw_mgr/isp_hw/ife_csid_hw/cam_ife_csid_hw_ver2.h
@@ -552,7 +552,7 @@ struct cam_ife_csid_ver2_common_reg_info {
 struct cam_ife_csid_secure_info {
 	uint32_t phy_sel;
 	uint32_t lane_cfg;
-	uint32_t vc_mask;
+	uint64_t vc_mask;
 	uint32_t csid_hw_idx_mask;
 	uint32_t cdm_hw_idx_mask;
 };
